[loader] workaround for nylas/N1#436
Because the key for the cached message is taken with the original body of the message
which is "" for Facebook messages, the messages have the same key.

This change avoids resetting the cache for each message that has a null body.
Unfortunately `MessageLoaderHeader` uses the private variable `_subscriptions`
to avoid resetting the cache and only signals the callback that matches the
current message.